EBView Code Cleaning

+++ ElCanari/EBView/EBView-key-down.swift
@@ -11,20 +11,20 @@ extension EBView {
   //····················································································································
 
   override func keyDown (with inEvent: NSEvent) {
-    let amount : CGFloat = inEvent.modifierFlags.contains (.shift)
+    let amount : Int = inEvent.modifierFlags.contains (.shift)
       ? self.shiftArrowKeyMagnitude
       : self.arrowKeyMagnitude
     ;
     for character in (inEvent.characters ?? "").unicodeScalars {
       switch (character) {
       case NSEvent.SpecialKey.upArrow.unicodeScalar :
-        _ = self.wantsToTranslateSelection (byX: 0.0, byY:amount)
+        _ = self.wantsToTranslateSelection (byX: 0, byY:amount)
       case NSEvent.SpecialKey.downArrow.unicodeScalar :
-        _ = self.wantsToTranslateSelection (byX: 0.0, byY:-amount)
+        _ = self.wantsToTranslateSelection (byX: 0, byY:-amount)
       case NSEvent.SpecialKey.leftArrow.unicodeScalar :
-        _ = self.wantsToTranslateSelection (byX: -amount, byY:0.0)
+        _ = self.wantsToTranslateSelection (byX: -amount, byY: 0)
       case NSEvent.SpecialKey.rightArrow.unicodeScalar :
-        _ = self.wantsToTranslateSelection (byX: amount, byY:0.0)
+        _ = self.wantsToTranslateSelection (byX: amount, byY: 0)
       case NSEvent.SpecialKey.deleteForward.unicodeScalar, NSEvent.SpecialKey.delete.unicodeScalar :
         self.deleteSelection ()
       default :
@@ -35,7 +35,7 @@ extension EBView {
 
   //····················································································································
 
-  private func wantsToTranslateSelection (byX inDx: CGFloat, byY inDy: CGFloat) -> Bool {
+  private func wantsToTranslateSelection (byX inDx: Int, byY inDy: Int) -> Bool {
     var accepted = true
     for object in self.viewController?.selectedGraphicObjectSet ?? [] {
       if !object.acceptToTranslate (xBy: inDx, yBy:inDy) {
